requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Environment ───────────────────────────────────────────────
python-dotenv==1.0.1

# ── Database ──────────────────────────────────────────────────
supabase==2.10.0

# ── Redis (Upstash) ───────────────────────────────────────────
upstash-redis==1.3.0

# ── Vector DB ─────────────────────────────────────────────────
chromadb==0.6.3

# ── AI / LLM ──────────────────────────────────────────────────
groq==0.13.1
langchain-groq==0.2.4
langchain==0.3.14

# ── Embeddings ────────────────────────────────────────────────
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1

# ── Token Counting ────────────────────────────────────────────
tiktoken==0.8.0

# ── Scheduler ─────────────────────────────────────────────────
apscheduler==3.10.4

# ── Encryption ────────────────────────────────────────────────
cryptography==44.0.0

# ── Data Validation ───────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1

# ── HTTP Client ───────────────────────────────────────────────
httpx==0.28.1

# ── Utilities ─────────────────────────────────────────────────
numpy==1.26.4